core/network/src/main/java/com/truongdc/movie/core/network/MovieService.kt
/*
 * Designed and developed by 2024 truongdc21 (Dang Chi Truong)
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package com.truongdc.movie.core.network

import com.truongdc.movie.core.model.BaseResponse
import com.truongdc.movie.core.network.model.NetworkMovie
import retrofit2.http.GET
import retrofit2.http.Path
import retrofit2.http.Query

interface MovieService {

    @GET("movie/top_rated?")
    suspend fun fetchTopRateMovies(
        @Query("api_key") apiKey: String?,
        @Query("page") page: Int = 1,
    ): BaseResponse<List<NetworkMovie>>

    @GET("movie/{movieId}")
    suspend fun fetchMovieDetails(
        @Path("movieId") movieId: Int,
        @Query("api_key") apiKey: String?,
    ): NetworkMovie
}
